+++ src/status/connection-status.ts
@@ -36,6 +36,6 @@ export async function updateActiveConnectionStatusBarItem(
   activeConnectionStatusBarItem.command = 'neo4j.setActiveConnection'
 
   activeConnectionStatusBarItem.text = active
-    ? `$(database) ${active.scheme}://${active.username}@${active.host}:${active.port}`
+    ? `$(database) ${active.scheme}://${active.username}@${active.host}:${active.port}${active.database ? `/${active.database}` : ''}`
     : '$(database) No Active Neo4j Connection'
 }